Handle deadlock exceptions in the postgresql cache
We've seen this happen in deletes, but we'll handle it in writes too.
For deletes, just ignore the exception and proceed (if deleting things
from the cache wants to fail, that's fine, we'll get to it later).

Haven't seen it happen with writes, but since it's possible for writes
and deletes to conflict, it's possible the writes could draw the short
straw.  There we won't fail, we'll just roll back the transaction and
try again.